requirements.txt
numpy>=1.24
scipy>=1.11
scikit-learn>=1.3
torch>=2.1
transformers>=4.40
datasets>=2.18
faiss-cpu>=1.7.4
pillow>=10.0
imageio>=2.34
pydantic>=2.6
pyyaml>=6.0
joblib>=1.3
pytest>=8.0
